refactor: rename `Blacklist` to `Denylist`
