+++ .changeset/little-dragons-sell.md
@@ -1,0 +1,5 @@
+---
+"kiai.js": patch
+---
+
+Rename `Blacklist` to `Denylist`
