--- packages/kiai.js/src/handlers/Blacklist.ts
@@ -1,41 +1,0 @@
-import { Blacklist as APIBlacklist, Message } from "@buape/kiai-api-types"
-import { BaseHandler } from "."
-
-export class Blacklist extends BaseHandler {
-	async getBlacklists(guildId: string) {
-		const result = (await this._handler.request(
-			`/guild/${guildId}/blacklist`
-		)) as APIBlacklist[]
-		return result
-	}
-
-	async createBlacklist(guildId: string, data: APIBlacklist) {
-		const result = (await this._handler.request(
-			`/guild/${guildId}/blacklist`,
-			"POST",
-			{},
-			data
-		)) as Message
-		return result
-	}
-
-	async deleteAllBlacklists(guildId: string) {
-		const result = (await this._handler.request(
-			`/guild/${guildId}/blacklist`,
-			"DELETE"
-		)) as {
-			users: number
-			roles: number
-			channels: number
-		}
-		return result
-	}
-
-	async deleteBlacklist(guildId: string, data: APIBlacklist) {
-		const result = (await this._handler.request(
-			`/guild/${guildId}/blacklist/${data.type}/${data.id}`,
-			"DELETE"
-		)) as Message
-		return result
-	}
-}
